backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
openpyxl==3.1.2
Pillow==10.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0